requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.2.0
    # via httpx
apscheduler==3.10.4
    # via python-telegram-bot
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via python-telegram-bot
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot[job-queue]==20.7
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   apscheduler
    #   python-telegram-bot
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via apscheduler
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
tzlocal==5.2
    # via apscheduler
urllib3==2.1.0
    # via requests
